ams-backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
psycopg==3.3.3
psycopg-binary==3.3.3
sqlparse==0.5.5
tzdata==2026.1
gunicorn==21.2.0
